Updating FCEE for numpy2 compatibility.
RE:#1641

+++ src/natcap/invest/forest_carbon_edge_effect.py
@@ -583,13 +583,14 @@ def _aggregate_carbon_map(
     for poly_feat in target_aggregate_layer:
         poly_fid = poly_feat.GetFID()
         poly_feat.SetField(
-            'c_sum', serviceshed_stats[poly_fid]['sum'])
+            'c_sum', float(serviceshed_stats[poly_fid]['sum']))
         # calculates mean pixel value per ha in for each feature in AOI
         poly_geom = poly_feat.GetGeometryRef()
         poly_area_ha = poly_geom.GetArea() / 1e4  # converts m^2 to hectare
         poly_geom = None
         poly_feat.SetField(
-            'c_ha_mean', serviceshed_stats[poly_fid]['sum']/poly_area_ha)
+            'c_ha_mean',
+            float(serviceshed_stats[poly_fid]['sum'] / poly_area_ha))
 
         target_aggregate_layer.SetFeature(poly_feat)
     target_aggregate_layer.CommitTransaction()
